# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tooling Replacement Workflow",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/excel/preview": {
      "post": {
        "tags": ["Excel"],
        "summary": "Preview Excel",
        "operationId": "preview_excel_api_excel_preview_post",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": ["create", "upsert"],
              "type": "string",
              "default": "create",
              "title": "Mode"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_excel_api_excel_preview_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/excel/import": {
      "post": {
        "tags": ["Excel"],
        "summary": "Import Excel",
        "operationId": "import_excel_api_excel_import_post",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": ["create", "upsert"],
              "type": "string",
              "default": "create",
              "title": "Mode"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_excel_api_excel_import_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/excel/export": {
      "get": {
        "tags": ["Excel"],
        "summary": "Export Excel",
        "operationId": "export_excel_api_excel_export_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_excel_api_excel_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": ["file"],
        "title": "Body_import_excel_api_excel_import_post"
      },
      "Body_preview_excel_api_excel_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": ["file"],
        "title": "Body_preview_excel_api_excel_preview_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportIssue": {
        "properties": {
          "sheet": {
            "type": "string",
            "title": "Sheet"
          },
          "row": {
            "type": "integer",
            "title": "Row"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": ["sheet", "row", "message"],
        "title": "ImportIssue"
      },
      "ImportReport": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "created_projects": {
            "type": "integer",
            "title": "Created Projects",
            "default": 0
          },
          "updated_projects": {
            "type": "integer",
            "title": "Updated Projects",
            "default": 0
          },
          "action_rows": {
            "type": "integer",
            "title": "Action Rows",
            "default": 0
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ImportIssue"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          },
          "project_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Project Ids",
            "default": []
          }
        },
        "type": "object",
        "required": ["valid"],
        "title": "ImportReport"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}